adjust variable names, move to internal, move testdata
