+++ processor/logtypedetectionprocessor/internal/fingerprint/fingerprint.go
