+++ processor/logtypedetectionprocessor/internal/fingerprint/fingerprint_bench_test.go
@@ -21,11 +21,11 @@ import (
 )
 
 func BenchmarkFingerprintJSONLogs(b *testing.B) {
-	benchmarkFingerprintCorpus(b, "../testdata/jsonLogs.csv")
+	benchmarkFingerprintCorpus(b, "testdata/jsonLogs.csv")
 }
 
 func BenchmarkFingerprintXMLLogs(b *testing.B) {
-	benchmarkFingerprintCorpus(b, "../testdata/xmlLogs.csv")
+	benchmarkFingerprintCorpus(b, "testdata/xmlLogs.csv")
 }
 
 func benchmarkFingerprintCorpus(b *testing.B, path string) {
