+++ processor/logtypedetectionprocessor/internal/fingerprint/fingerprint_json.go
