+++ processor/logtypedetectionprocessor/internal/fingerprint/fingerprint_xml.go
